+++ packages/contracts/contracts/core/permission/PermissionLib.sol
@@ -9,7 +9,6 @@ library PermissionLib {
     enum Operation {
         Grant,
         Revoke,
-        Freeze,
         GrantWithOracle
     }